Refactor MPLAPACK test code using FEM for string handling and I/O
Major improvements using FABLE's FEM (Fortran EMulation) module for
better code quality, readability, and multi-precision I/O support.

== String handling improvements ==
- Replace const char* with fem::str_cref for function parameters
- Replace char arrays with fem::str<N> for local string variables
- Use FEM string slicing: path(1,3), line(i,i), intstr(k,k)
- Update Mlsame() calls: Mlsame(pack, "N") -> Mlsame(pack.elems(), "N")

== FORMAT statements reorganization ==
- Define format strings as static const char *format_NNNN
- Align format labels with original FORTRAN code order (9999, 9998, ...)
- Group related formats together with comments (GQR/GRQ/LSE/GLM/GSVD/CSD)
- Move format definitions to function top for clarity

== Write output improvements ==
- Remove sprintnum_short(buf, result) workaround
- Use FEM write directly: write(nounit, format_NNNN), ..., result
- FEM internally uses snprintnum_short for multi-precision output
- FORMAT specifiers are parsed but actual output uses snprintnum_short

== Read input improvements ==
- Remove C++ iostream/istringstream/regex dependencies
- Use FEM common_read: read(nin, "(a80)"), line
- FEM modification enables direct multi-precision value reading
- Simplify integer parsing with FEM string operations
- Add proper read_end exception handling with goto statements

== File renames ==
- chkxer.cpp -> Chkxer.cpp (consistent naming)
- xlaenv.cpp -> Mxlaenv.cpp (MPLAPACK prefix)
- Remove Rlacsg.cpp, Clacsg.cpp (unused)

== Memory allocation improvements ==
- Replace stack arrays with std::make_unique<T[]> for large matrices
- Ensures proper heap allocation for test matrices (50x50)

== Numeric literal cleanup ==
- 0.3e0 -> 0.3 (simpler notation)
- Remove unnecessary REAL() casts in max() calls
- Fix max({...}) brace usage for two-argument cases

== Bug fixes ==
- Fix array index: ldv -> ldu, ldb -> lda in several routines
- Add Rlabad(safmin, safmax) call in Cchkgg
- Fix rmagn array indexing with parentheses

This refactoring improves code maintainability by aligning more closely
with the original FORTRAN structure while leveraging FEM for proper
multi-precision arithmetic I/O.

+++ mplapack/test/eig/common/Rsxt1.cpp
@@ -43,8 +43,6 @@ using fem::common;
 #include <mplapack_matgen.h>
 #include <mplapack_eig.h>
 
-#include <mplapack_debug.h>
-
 REAL Rsxt1(INTEGER const ijob, REAL *d1, INTEGER const n1, REAL *d2, INTEGER const n2, REAL const abstol, REAL const ulp, REAL const unfl) {
     REAL return_value = 0.0;
     const REAL zero = 0.0;
@@ -65,12 +63,12 @@ REAL Rsxt1(INTEGER const ijob, REAL *d1, INTEGER const n1, REAL *d2, INTEGER con
         if (j == 1) {
             temp2 = abs(d2[j - 1] - d1[i - 1]);
             if (ijob == 2) {
-                temp2 = temp2 / max(unfl, REAL(abstol + ulp * abs(d1[i - 1])));
+                temp2 = temp2 / max(unfl, abstol + ulp * abs(d1[i - 1]));
             }
         } else {
             temp2 = min(abs(d2[j - 1] - d1[i - 1]), abs(d1[i - 1] - d2[(j - 1) - 1]));
             if (ijob == 2) {
-                temp2 = temp2 / max(unfl, REAL(abstol + ulp * abs(d1[i - 1])));
+                temp2 = temp2 / max(unfl, abstol + ulp * abs(d1[i - 1]));
             }
         }
         temp1 = max(temp1, temp2);
